+++ src/controller/display/PathDisplay.module.scss
@@ -1,0 +1,13 @@
+.wrapper {
+  display: grid;
+  grid-template-columns: subgrid;
+  grid-column: 1/3;
+  align-items: center;
+  .label {
+    margin-right: 10px;
+    grid-column: 1/2;
+  }
+  .content {
+    grid-column: 2/3;
+  }
+}
